fix: parse HTTP without connection close
