+++ Cargo.toml
@@ -62,6 +62,7 @@ chrono = "0.4.41"
 dialoguer = "0.12.0"
 once_cell = "1.21.3"
 papaya = "0.2.3"
+httparse = "1.8"
 
 [dev-dependencies]
 tokio-test = "0.4"
@@ -72,3 +73,7 @@ lto = true
 codegen-units = 1
 panic = "abort"
 strip = true
+
+[[bin]]
+name = "shadowmap-server"
+path = "src/server.rs"
